IMS-VT: Change call data usage parameters
Change call data usage methods to take parameter of long type
instead of int

CRs-Fixed: 747694

+++ src/java/com/android/ims/internal/ImsVideoCallProvider.java
@@ -222,7 +222,7 @@ public abstract class ImsVideoCallProvider {
     }
 
     /** @see Connection.VideoProvider#changeCallDataUsage */
-    public void changeCallDataUsage(int dataUsage) {
+    public void changeCallDataUsage(long dataUsage) {
         if (mCallback != null) {
             try {
                 mCallback.changeCallDataUsage(dataUsage);
